+++ src/index.ts
@@ -1,5 +1,5 @@
 import { Err, Ok, Result } from "./Result"
 
-type AsyncResult<T, E extends Error> = Promise<Result<T, E>>
+type AsyncResult<T, E = Error> = Promise<Result<T, E>>
 
 export { Result, Err, Ok, AsyncResult }