+++ CHANGELOG.md
@@ -1,5 +1,95 @@
 # @flagg2/result
 
+## 1.2.1
+
+### Patch Changes
+
+-  remove deprecated apis
+
+## 1.2.0
+
+### Minor Changes
+
+-  merge fromAsync, fromSync, fromPromise into one function
+
+## 1.1.2
+
+### Patch Changes
+
+-  fromAsync -> catchFrom
+
+## 1.1.1
+
+### Patch Changes
+
+-  Add fromAsync
+
+## 1.1.0
+
+### Minor Changes
+
+-  upgrade typings
+
+## 1.0.17
+
+### Patch Changes
+
+-  x
+
+## 1.0.16
+
+### Patch Changes
+
+-  add andThen
+
+## 1.0.15
+
+### Patch Changes
+
+-  fix fromPromise
+
+## 1.0.14
+
+### Patch Changes
+
+-  p
+
+## 1.0.13
+
+### Patch Changes
+
+-  add default
+
+## 1.0.12
+
+### Patch Changes
+
+-  add default
+
+## 1.0.11
+
+### Patch Changes
+
+-  remove tovalue
+
+## 1.0.10
+
+### Patch Changes
+
+-  add toValue
+
+## 1.0.9
+
+### Patch Changes
+
+-  add default to async result
+
+## 1.0.8
+
+### Patch Changes
+
+-  add async result export
+
 ## 1.0.7
 
 ### Patch Changes